+++ src/spring-cloud/azext_spring_cloud/_transformers.py
@@ -34,8 +34,10 @@ def transform_app_table_output(result):
         item['Persistent Storage'] = "{}/{} Gb".format(
             persistentStorage['usedInGb'], persistentStorage['sizeInGb']) if persistentStorage['sizeInGb'] else "-"
 
-        if item['properties']['activeDeployment']:
+        if 'activeDeployment' in item['properties']:
             instances = item['properties']['activeDeployment']['properties']['instances']
+            if instances is None:
+                instances = []
             up_number = len(
                 [x for x in instances if x['discoveryStatus'] == 'UP'])
             item['Discovery Status'] = "UP( {} ), DOWN( {} )".format(
@@ -57,6 +59,8 @@ def transform_spring_cloud_deployment_output(result):
         item['App Name'] = item['properties']["appName"]
         item['State'] = item['properties']['provisioningState']
         instances = item['properties']['instances']
+        if instances is None:
+            instances = []
         up_number = len([x for x in instances if x['discoveryStatus'] == 'UP'])
         item['Discovery Status'] = "UP( {} ), DOWN( {} )".format(
             up_number, len(instances) - up_number)