+++ src/aks-preview/azext_aks_preview/_help.py
@@ -192,7 +192,7 @@ helps['aks create'] = """
           text: az aks create -g MyResourceGroup -n MyManagedCluster --kubernetes-version 1.13.9 --node-count 3 --enable-cluster-autoscaler --min-count 1 --max-count 5
         - name: Create a kubernetes cluster with k8s 1.13.9 but use vmas.
           text: az aks create -g MyResourceGroup -n MyManagedCluster --kubernetes-version 1.13.9 --vm-set-type AvailabilitySet
-        - name: Create a kubernetes cluster with default kubernetes vesrion, default SKU load balancer(basic) and default vm set type(AvailabilitySet).
+        - name: Create a kubernetes cluster with default kubernetes vesrion, default SKU load balancer(standard) and default vm set type(VirtualMachineScaleSets).
           text: az aks create -g MyResourceGroup -n MyManagedCluster
         - name: Create a kubernetes cluster with standard SKU load balancer and two AKS created IPs for the load balancer outbound connection usage.
           text: az aks create -g MyResourceGroup -n MyManagedCluster --load-balancer-managed-outbound-ip-count 2
@@ -360,7 +360,10 @@ helps['aks nodepool add'] = """
           long-summary: If not specified, defaults to 110, or 30 for advanced networking configurations.
         - name: --node-zones
           type: string array
-          short-summary: (PREVIEW) Availability zones where agent nodes will be placed.
+          short-summary: (will be deprecated, use --zones) Availability zones where agent nodes will be placed.
+        - name: --zones -z
+          type: string array
+          short-summary: Availability zones where agent nodes will be placed.
         - name: --vnet-subnet-id
           type: string
           short-summary: The ID of a subnet in an existing VNet into which to deploy the cluster.