+++ src/hack/azext_hack/__init__.py
@@ -3,6 +3,8 @@
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
 from azure.cli.core import AzCommandsLoader
+from azure.cli.core.commands.parameters import get_location_type
+import azext_hack._help  # pylint: disable=unused-import
 from ._validators import validate_name
 
 
@@ -15,14 +17,14 @@ class HackExtCommandLoader(AzCommandsLoader):
 
     def load_command_table(self, _):
         with self.command_group('hack') as g:
-            g.custom_command('up', 'hack_up')
+            g.custom_command('create', 'create_hack')
         return self.command_table
 
     def load_arguments(self, _):
-        with self.argument_context('hack up') as c:
+        with self.argument_context('hack create') as c:
             c.argument('name',
                        options_list=['--name', '-n'],
-                       help='Name of resources',
+                       help='Base name of resources; random charagers will be appended',
                        validator=validate_name,
                        type=str.lower)
             c.argument('database',
@@ -39,9 +41,10 @@ class HackExtCommandLoader(AzCommandsLoader):
                        type=str.lower)
             c.argument('ai',
                        help='Enable Azure Cognitive Services',
-                       options_list=['--ai', '-ai'],
+                       options_list=['--ai'],
                        default=None,
                        action='store_true')
+            c.argument(get_location_type(self.cli_ctx))
 
 
 COMMAND_LOADER_CLS = HackExtCommandLoader